requirements.txt
discord.py
psycopg2-binary
openpyxl
gspread
google-auth